API/requirements.txt
pandas
numpy
scikit-learn
flask
psycopg2-binary